backend/requirements.txt
fastapi[all]==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.6.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
pdfplumber==0.10.3
PyMuPDF==1.23.16
unstructured==0.12.3
langchain==0.1.17
langchain-community==0.0.36
tiktoken==0.6.0
pinecone-client==3.2.2
paypalrestsdk==1.13.3
schedule==1.2.0

# Email et réinitialisation de mot de passe
fastapi-mail==1.4.1
itsdangerous==2.1.2
jinja2==3.1.3
sentence-transformers==2.7.0